diagnonet-backend/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Machine Learning dependencies
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# BioGPT and Hugging Face integration (optimized versions)
transformers>=4.35.0
torch>=2.1.0
tokenizers>=0.15.0
protobuf>=3.20.0
sacremoses>=0.0.53

# Image processing and file handling
Pillow>=10.0.0
python-multipart>=0.0.6
